requirements.txt
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
transformers
streamlit
numpy
pillow
matplotlib
jupyter
datasets